Add Latency Sync support for any display resolution

+++ src/addons/display_commander/globals.hpp
@@ -118,8 +118,8 @@ extern std::atomic<int> s_fps_limiter_mode;
 // 0 = Sync to VBlank (current), 1 = Sync based on scanline (experimental)
 extern std::atomic<int> s_latent_sync_mode;
 
-// Scanline threshold for latent sync (0 to monitor height)
-extern std::atomic<int> s_scanline_threshold;
+// Scanline threshold ratio for latent sync (0.0 to 1.0, default 0.78 = 78%)
+extern std::atomic<float> s_scanline_threshold;
 extern std::atomic<int> s_scanline_window;
 
 // Performance stats (FPS/frametime) shared state
